Rename `DefaultMicrosoftIdentityAuthenticationDelegatingHandlerFactory`
Add `Default` prefix to the handler factory, make it `internal` and
split the interface and the implementation into separate files since the
naming scheme may make them harder to locate in the same file
